Tizen: Fix QCoreApplication::applicationDirPath()
Disable the code path which queries /proc/<pid>/exe for the
current executable path, as from a Q_OS_LINUX_TIZEN perspective,
this executable will be the launch_app binary. Instead we
get the application directory via the fallback, by
looking in argv[0], since this is set to the location of the
application binary.

Change-Id: I3132266b8fa83cfad968e5d04883d88a731c1a87
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>